chore: move disciplina to service and make it a class
